created new tests and new implementations for split up of classes
