+++ src/main/java/edu/ithaca/dragon/bank/AdminAPI.java
@@ -12,4 +12,6 @@ public interface AdminAPI {
 
     public void unfreezeAcct(String acctId);
 
+    public boolean getIsFrozen(String acctId);
+
 }
